Alternate method for exercise 2-6
An alternate method for finding the larger and smaller of two numbers, while only using concepts presented in the second chapter, is to compute the average, average difference, and the remainder of the division of the difference and 2 (this is because of loss of precision with integer division). Then, the larger number is the sum of all three, while the smaller number is the difference between the average and the average difference.